ams: Refactor workspace structure
This commit restructures the project to rename the `ams-core` crate to
just `ams`. This change is to make the core functionality of the project
more akin to a library crate that can be used by other applications.

The overall intent is to have the `ams` crate serve as the core
functionality of the project, while other crates can be built on top of
it to provide the different user interfaces. The current plan is to have
a TUI application, but who knows what the future holds.

Not only was the `ams-core` crate removed, but so was the `ams-tui`
crate. `ams-tui` crate was originally going to be a library crate
containing the different widgets that make up the TUI `ams` application.
However after an initial design of the TUI, only a few different widgets
should be needed, and does not justify its own crate. For the time
being, I am removing the crate as it is completely empty, however a
future commit will reintroduce it, and it will contain the TUI widgets
and binary application.

The other thing this commit does, which I can't be bothered to split
into its own commit (although I definitely should have), is to add an
initial, and simple, event stream to the `ams` crate to announce
different events happening that the consumer would not be otherwise
aware of. (e.g. A new incoming connection request, a peer disconnecting,
, a new message, etc). It also adds some initial methods for interacting
with the ams instance, such as initiating a connection, sending a
message, etc.